One component per file, and terser comments
Splits the panel into a file per component per the frontend convention, and
cuts every comment added here to two lines or fewer.
