+++ frontend/components/traces/trace-view/signal-events-panel/panel-body.tsx
@@ -8,19 +8,19 @@ import { shallow } from "zustand/shallow";
 
 import { laminarAgentStore } from "@/components/agent";
 import { useTraceViewStore } from "@/components/traces/trace-view/store";
-import { type TraceSignal } from "@/components/traces/trace-view/store/base";
 import { Button } from "@/components/ui/button";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { ElevatedSurface } from "@/components/ui/surface";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { Tabs, TabsContent, TabsList } from "@/components/ui/tabs";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
 import { useFeatureFlags } from "@/contexts/feature-flags-context.tsx";
 import { Feature } from "@/lib/features/features.ts";
 import { cn } from "@/lib/utils";
 
-import { TOOLTIP_DELAY_MS } from "./cluster-button";
-import SeverityIcon from "./severity-icon";
+import { TOOLTIP_DELAY_MS } from "./constants";
 import SignalDetails from "./signal-details";
+import SignalHeader from "./signal-header";
+import SignalTab from "./signal-tab";
 
 interface Props {
   traceId: string;
@@ -29,57 +29,6 @@ interface Props {
 
 const MIN_BODY_HEIGHT = 120;
 const MAX_BODY_HEIGHT = 320;
-
-/** The worst event's severity. With several events (a per-span signal) the header
- *  speaks for the whole card, and the card is as bad as its worst event. */
-const worstSeverity = (signal: TraceSignal) => signal.events.reduce((worst, e) => Math.max(worst, e.severity), 0);
-
-/**
- * The header when a trace has a single signal: severity, then the signal's name.
- * Severity is a glyph here rather than a `Critical` pill on the event's row — a
- * pill is the right shape for a list, and there is no list.
- */
-function SignalHeader({ signal }: { signal: TraceSignal }) {
-  return (
-    <div className="flex min-w-0 flex-1 items-center gap-1.5 pl-1">
-      {signal.events.length > 0 && <SeverityIcon severity={worstSeverity(signal)} />}
-      <span className="min-w-0 truncate text-xs font-medium">{signal.signalName}</span>
-    </div>
-  );
-}
-
-/**
- * One signal tab.
- *
- * Its own component because of the tooltip: several tabs in a panel this wide
- * truncate to a few characters, at which point the row stops naming anything.
- * The tooltip trigger is a WRAPPER, not the tab itself — both primitives own a
- * `data-state` (`active`/`inactive` vs `closed`/`delayed-open`) and
- * `Tabs.Trigger` spreads incoming props after its own attributes, so `asChild`
- * onto the tab overwrites `active` with `closed` and every `data-[state=active]:`
- * rule in the trigger's base classes stops matching.
- */
-function SignalTab({ signal }: { signal: TraceSignal }) {
-  const trigger = (
-    <TabsTrigger value={signal.signalId} className="w-full min-w-0">
-      {/* The tab row REPLACES the header, so without this the severity glyph
-          disappears exactly when there is more than one severity to tell apart. */}
-      {signal.events.length > 0 && <SeverityIcon bare severity={worstSeverity(signal)} />}
-      <span className="min-w-0 truncate">{signal.signalName}</span>
-    </TabsTrigger>
-  );
-
-  return (
-    <Tooltip delayDuration={TOOLTIP_DELAY_MS}>
-      <TooltipTrigger asChild>
-        <span className="flex min-w-0 flex-1">{trigger}</span>
-      </TooltipTrigger>
-      <TooltipPortal>
-        <TooltipContent side="bottom">{signal.signalName}</TooltipContent>
-      </TooltipPortal>
-    </Tooltip>
-  );
-}
 
 export default function PanelBody({ traceId, onClose }: Props) {
   const contentRef = useRef<HTMLDivElement>(null);
@@ -183,24 +132,21 @@ export default function PanelBody({ traceId, onClose }: Props) {
       ) : (
         <Tabs value={effectiveTabId} onValueChange={setActiveSignalTabId} className="flex flex-col gap-0">
           <TooltipProvider delayDuration={TOOLTIP_DELAY_MS}>
-            {/* The leading inset matches the trailing one under tabs — the tab
-                list is a filled pill with its own inset, so it wants to start
-                where the close button ends. The single-signal header is bare
-                text and keeps the wider inset. */}
+            {/* The tab list is a filled pill with its own inset, so under tabs the
+                leading inset matches the trailing one; bare header text keeps the
+                wider one. */}
             <div className={cn("flex shrink-0 items-center justify-between gap-2 p-1.5", isSingleSignal && "pl-2")}>
               {isSingleSignal && activeSignal ? (
                 <SignalHeader signal={activeSignal} />
               ) : (
                 <TabsList size="sm" className="min-w-0 flex-1 justify-start">
                   {traceSignals.map((signal) => (
-                    <SignalTab key={signal.signalId} signal={signal} />
+                    <SignalTab key={signal.signalId} signal={signal} active={signal.signalId === effectiveTabId} />
                   ))}
                 </TabsList>
               )}
-              {/* "Open in AI Chat" lost its actions row when that row collapsed
-                  into the header, so it becomes an icon beside the close button:
-                  it is feature-flagged and usually absent, and a second worded
-                  button would outweigh the signal's own name. */}
+              {/* An icon, not a worded button: it is feature-flagged and usually
+                  absent, and a second label would outweigh the signal's own name. */}
               {featureFlags[Feature.AGENT] && (
                 <Tooltip>
                   <TooltipTrigger asChild>
