+++ frontend/components/traces/trace-view/signal-events-panel/signal-event.tsx
@@ -1,0 +1,92 @@
+"use client";
+
+import { useEffect, useMemo, useRef } from "react";
+
+import { type SchemaField } from "@/components/signals/utils";
+import { type SpanReferenceCallbacks } from "@/components/traces/trace-view/span-reference";
+import { type TraceSignalEvent } from "@/components/traces/trace-view/store/base";
+import { cn } from "@/lib/utils";
+
+import ClusterButton from "./cluster-button";
+import OpenInSignalsButton from "./open-in-signals-button";
+import PayloadValue from "./payload-value";
+
+function parsePayload(payload: string): Record<string, unknown> {
+  try {
+    const result = JSON.parse(payload);
+    if (result == null || typeof result !== "object" || Array.isArray(result)) {
+      return {};
+    }
+    return result;
+  } catch {
+    return {};
+  }
+}
+
+interface Props {
+  event: TraceSignalEvent;
+  projectId: string;
+  signalId: string;
+  traceId: string;
+  validFields: SchemaField[];
+  spanRefCallbacks?: SpanReferenceCallbacks;
+  highlighted?: boolean;
+}
+
+/** One event, with no card around it: a signal produces one event per trace in
+ *  all but a handful of cases, so the box bordered the panel's only content. */
+export default function SignalEvent({
+  event,
+  projectId,
+  signalId,
+  traceId,
+  validFields,
+  spanRefCallbacks,
+  highlighted,
+}: Props) {
+  const parsed = useMemo(() => parsePayload(event.payload), [event.payload]);
+
+  const ref = useRef<HTMLDivElement>(null);
+  useEffect(() => {
+    if (highlighted) {
+      ref.current?.scrollIntoView({ block: "nearest" });
+    }
+  }, [highlighted]);
+
+  return (
+    <div
+      ref={ref}
+      // A deep-linked event still has to be findable, so the highlight is a rule
+      // down the left edge rather than a ring around a box that is no longer drawn.
+      className={cn("flex min-w-0 flex-col", highlighted && "border-l-2 border-signal/60")}
+    >
+      {/* The SIDE inset is the fields' one: this row and the payload are two blocks
+          in one column, and a column has one left edge. Only the vertical is ours. */}
+      <div className="flex min-w-0 items-center gap-1.5 px-3 py-2">
+        {event.leafClusters.length > 0 ? (
+          event.leafClusters.map((cluster) => (
+            <ClusterButton
+              key={cluster.id}
+              shrinkable
+              cluster={cluster}
+              href={`/project/${projectId}/signals/${signalId}?clusterId=${cluster.id}&traceId=${traceId}&eventId=${event.id}`}
+            />
+          ))
+        ) : (
+          <OpenInSignalsButton href={`/project/${projectId}/signals/${signalId}?traceId=${traceId}`} />
+        )}
+      </div>
+
+      <div className="flex min-w-0 flex-col gap-2.5 px-3 pb-0.5">
+        {validFields.map((field) => (
+          <div key={field.name} className="flex flex-col gap-0.5">
+            <div className="text-xs font-medium text-signal-key">{field.name}</div>
+            <div className="text-sm leading-relaxed text-secondary-foreground">
+              <PayloadValue value={parsed[field.name]} field={field} spanRefCallbacks={spanRefCallbacks} />
+            </div>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+}
